backend/requirements.txt
# ============================================================================
# PRODUCTION DEPENDENCIES
# ============================================================================

# Core FastAPI and web framework
fastapi>=0.104.0
fastapi-cache2[redis]==0.2.2
uvicorn[standard]>=0.24.0

# Database and ORM
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Authentication and Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=3.2.0,<4.0.0
python-multipart>=0.0.6

# Configuration and Environment
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# HTTP Client and Networking
httpx>=0.25.0
requests>=2.31.0

# Rate Limiting and Middleware
slowapi>=0.1.9
redis>=4.2.0,<5.0.0

# Data Processing and Analysis
numpy==1.26.4
pandas==2.2.2
ta

# Financial Data APIs
baostock>=0.8.8
yfinance>=0.2.18
akshare==1.17.35
yahoo_fin>=0.8.9
requests_html>=0.10.0

# Task Scheduling
apscheduler>=3.10.0

# Logging and Monitoring
loguru>=0.7.0
colorlog>=6.7.0

# Email Support
email-validator>=2.1.0

# File Handling
aiofiles>=23.2.0

# System Monitoring
psutil>=5.9.0

# ============================================================================
# NOTE: Development dependencies moved to requirements-dev.txt
# For development: pip install -r requirements-dev.txt
# ============================================================================
